Emit events in more sensible order
If you subscribe to "state:offline" and "reconnectionPending":

  mobidy.on("state:offline", console.log);
  mobidy.on("reconnectionPending", console.log);

you will currently get these events in the reverse order:

  reconnectionPending
  state:offline

This defies expectations: I am trying to model the state of my application such that I can only have pending reconnections while I am offline, and this is making it hard to do that.

The reason this occurs is that Mopidy.js uses EventEmitter to allow external listeners to be added and also for its internal listeners. When "websocket:close" is emitted, "state:offline" is emitted. The listeners for "state:offline" are:

    [
      Mopidy.js `_reconnect` method,
      My state:offline listener
    ]

The Mopidy.js `_reconnect` handler in term emits "reconnectionPending". The only event listener for that is mine. Even though we're already in the middle of emitting "state:offline" and haven't yet gotten to emitting it to the consumer, we will synchronously emit "reconnectionPending". In other words, the order of events is:

1. "state:offline" is emitted
1. "state:offline": 1st event handler executed (`_reconnect` called)
1. `_reconnect` emits "reconnectionsPending"
1. My "reconnectionsPending" listener is executed
1. "state:offline": 2nd event handler executed (my listener is called)

To fix this, this PR makes it so that all the internal listeners of events execute asynchronously using `setImmediate`. It is safe to make the assumption that all listeners will be able to handle that, because we're already asynchronously responding to websock events. No consuming code can assume it is synchronously doing something.

The new behavior is now, when an event is emitted, for all listeners:
  - if an internal listener (e.g. `_reconnect`), queue up asynchronously
  - if an external listener execute synchronously (the listener is free to delegate with `setImmediate` in the same way)
  - the asynchronous internal listeners execute on next tick, possibly emitting more events

I've had to modify some of the existing tests to handle the fact that Jest's fake timers need to run after an event is emitted, so that `setImmediate` executes.

I have added a test that demonstrates that "state:offline" fires before "reconnectionPending".